Read metadata

Read metadata

Usage

read_metadata(
  dataset_id = "s_wfs_alkis_bezirk",
  service_type = "WFS",
  encoding = "Windows-1252",
  debug = TRUE
)

Arguments

dataset_id

id of wfs dataset (default: "s_wfs_alkis_bezirk"). Needs to checked out manually as described below:

Go to: https://fbinter.stadt-berlin.de/fb/

Click on a "WFS" or "WMS" Dataset

Read the content of Rechneraddresse in the opened tab. The basename of the url is the required dataset id!

service_type

either "WFS" or "WMS" (default: "WFS")

encoding

encoding of metadata FIS-Broker site (default: "Windows-1252")

debug

print debug messages (default: TRUE)

Value

tibble with metadata for provided dataset_id

See also

Examples

### One Dataset
berlin_bezirke_metadata <- read_metadata(dataset_id = "s_wfs_alkis_bezirk")
#> Importing WFS metadata for dataset_id 's_wfs_alkis_bezirk' from FIS-Broker ... ok. (0.10s) 
berlin_bezirke_metadata
#> # A tibble: 17 × 4
#>    title_id title_name            parameter              value                  
#>       <int> <chr>                 <chr>                  <chr>                  
#>  1        1 ALKIS Berlin Gemeinde Kurzbeschreibung       "Bezirke von Berlin"   
#>  2        1 ALKIS Berlin Gemeinde Koordinatensysteme     "EPSG:25833, EPSG:4258"
#>  3        1 ALKIS Berlin Gemeinde Ausdehnung             "EPSG:25833: 369097,85…
#>  4        1 ALKIS Berlin Gemeinde Ausdehnung             "EPSG:4258: 13,08, 52,…
#>  5        1 ALKIS Berlin Gemeinde Nutzungsbedingungen    "Für die Nutzung der D…
#>  6        1 ALKIS Berlin Gemeinde Zugriffsbeschränkungen "Es gelten keine Zugri…
#>  7        2 Technische Angaben    Rechneradresse         "https://fbinter.stadt…
#>  8        2 Technische Angaben    Standard               "WFS"                  
#>  9        2 Technische Angaben    Version                "1.0.0"                
#> 10        2 Technische Angaben    Version                "1.1.0"                
#> 11        2 Technische Angaben    Version                "2.0.0"                
#> 12        2 Technische Angaben    Operationen            "GetCapabilities"      
#> 13        2 Technische Angaben    Operationen            "DescribeFeatureType"  
#> 14        2 Technische Angaben    Operationen            "GetFeature"           
#> 15        2 Technische Angaben    Ausgabeformate         "text/xml; subtype=gml…
#> 16        2 Technische Angaben    Ausgabeformate         "text/xml; subtype=gml…
#> 17        2 Technische Angaben    Ausgabeformate         "text/xml; subtype=gml…

### Multiple
ids_wfs <- readLines(system.file("extdata/ids_wfs.txt", 
package = "kwb.fisbroker"))

wfs_meta_list <- setNames(lapply(ids_wfs, function(id) {
kwb.fisbroker::read_metadata(id)}),
ids_wfs)
#> Importing WFS metadata for dataset_id 's03_11_3LRP2018_2025' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 'sach_nutz2020_nutzsa' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 'sach_nutz2015_nutzsa' from FIS-Broker ... ok. (0.06s) 
#> Importing WFS metadata for dataset_id 's08_07_2tehganlagen' from FIS-Broker ... ok. (0.04s) 
#> Importing WFS metadata for dataset_id 's03_12_2emissionen' from FIS-Broker ... ok. (0.04s) 
#> Importing WFS metadata for dataset_id 's_messpunkte' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's02_08fisch2013' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 'sach_nutz2020_nutzsa' from FIS-Broker ... ok. (0.04s) 
#> Importing WFS metadata for dataset_id 'sach_nutz2015_nutzsa' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_05_09gruenvol2020' from FIS-Broker ... ok. (0.04s) 
#> Importing WFS metadata for dataset_id 'co2veg_block' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 'co2veg_strasse' from FIS-Broker ... ok. (0.04s) 
#> Importing WFS metadata for dataset_id 's_04_08_2lniederschl_bl_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_08_2lniederschl_fl_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_08_2lniederschl_str_8110' from FIS-Broker ... ok. (0.04s) 
#> Importing WFS metadata for dataset_id 's_04_08_3lniederschl_bl_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_08_3lniederschl_fl_8110' from FIS-Broker ... ok. (0.04s) 
#> Importing WFS metadata for dataset_id 's_04_08_3lniederschl_str_8110' from FIS-Broker ... ok. (0.04s) 
#> Importing WFS metadata for dataset_id 's_04_02_1ltempmittel_bl_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_02_1ltempmittel_fl_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_02_1ltempmittel_str_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_02_2ltempmittel_bl_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_02_2ltempmittel_str_8110' from FIS-Broker ... ok. (0.04s) 
#> Importing WFS metadata for dataset_id 's_04_02_3ltempmittel_bl_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_02_3ltempmittel_fl_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_02_3ltempmittel_str_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_02_5ltempmittel_bl_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_02_5ltempmittel_fl_8110' from FIS-Broker ... ok. (0.05s) 
#> Importing WFS metadata for dataset_id 's_04_02_5ltempmittel_str_8110' from FIS-Broker ... ok. (0.04s) 
#> Importing WFS metadata for dataset_id 's03_11_3LRP2018_2025' from FIS-Broker ... ok. (0.05s) 

wfs_meta <- dplyr::bind_rows(wfs_meta_list, .id = "id_wfs")

wfs_meta
#> # A tibble: 495 × 5
#>    id_wfs               title_id title_name                        param…¹ value
#>    <chr>                   <int> <chr>                             <chr>   <chr>
#>  1 s03_11_3LRP2018_2025        1 Luftreinhalteplan 2018-2025 - Sz… Kurzbe… "Wir…
#>  2 s03_11_3LRP2018_2025        1 Luftreinhalteplan 2018-2025 - Sz… Koordi… "EPS…
#>  3 s03_11_3LRP2018_2025        1 Luftreinhalteplan 2018-2025 - Sz… Ausdeh… "EPS…
#>  4 s03_11_3LRP2018_2025        1 Luftreinhalteplan 2018-2025 - Sz… Nutzun… "Für…
#>  5 s03_11_3LRP2018_2025        1 Luftreinhalteplan 2018-2025 - Sz… Zugrif… "Es …
#>  6 s03_11_3LRP2018_2025        2 Technische Angaben                Rechne… "htt…
#>  7 s03_11_3LRP2018_2025        2 Technische Angaben                Standa… "WFS"
#>  8 s03_11_3LRP2018_2025        2 Technische Angaben                Version "1.0…
#>  9 s03_11_3LRP2018_2025        2 Technische Angaben                Version "1.1…
#> 10 s03_11_3LRP2018_2025        2 Technische Angaben                Version "2.0…
#> # … with 485 more rows, and abbreviated variable name ¹​parameter