Read metadata

Read metadata

Usage

read_metadata(
  dataset_id = "s_wfs_alkis_bezirk",
  service_type = "WFS",
  encoding = "Windows-1252",
  dbg = TRUE,
  url = NULL
)

Arguments

dataset_id

id of wfs dataset (default: "s_wfs_alkis_bezirk"). Needs to checked out manually as described below:

Go to: https://fbinter.stadt-berlin.de/fb/

Click on a "WFS" or "WMS" Dataset

Read the content of Rechneraddresse in the opened tab. The basename of the url is the required dataset id!

service_type

either "WFS" or "WMS" (default: "WFS")

encoding

encoding of metadata FIS-Broker site (default: "Windows-1252")

dbg

print debug messages (default: TRUE)

url

optional. URL to web resource from which to load metadata. by default generated by the function

Value

tibble with metadata for provided dataset_id

See also

Examples

### One Dataset
berlin_bezirke_metadata <- read_metadata(dataset_id = "s_wfs_alkis_bezirk")
#> Getting HTML text from 'https://fbinter.stadt-be...ezirk@senstadt&type=WFS' ... ok. (0.23s) 
berlin_bezirke_metadata
#>                  section              parameter
#> 1  ALKIS Berlin Gemeinde       Kurzbeschreibung
#> 2  ALKIS Berlin Gemeinde     Koordinatensysteme
#> 3  ALKIS Berlin Gemeinde             Ausdehnung
#> 4  ALKIS Berlin Gemeinde             Ausdehnung
#> 5  ALKIS Berlin Gemeinde    Nutzungsbedingungen
#> 6  ALKIS Berlin Gemeinde Zugriffsbeschränkungen
#> 7     Technische Angaben         Rechneradresse
#> 8     Technische Angaben               Standard
#> 9     Technische Angaben                Version
#> 10    Technische Angaben                Version
#> 11    Technische Angaben                Version
#> 12    Technische Angaben            Operationen
#> 13    Technische Angaben            Operationen
#> 14    Technische Angaben            Operationen
#> 15    Technische Angaben         Ausgabeformate
#> 16    Technische Angaben         Ausgabeformate
#> 17    Technische Angaben         Ausgabeformate
#>                                                                                                                                                                                                                                                                  value
#> 1                                                                                                                                                                                                                                                   Bezirke von Berlin
#> 2                                                                                                                                                                                                                                                EPSG:25833, EPSG:4258
#> 3                                                                                                                                                                                                             EPSG:25833: 369097,85, 5799298,14, 416865,04, 5838236,21
#> 4                                                                                                                                                                                                                                EPSG:4258: 13,08, 52,33, 13,77, 52,69
#> 5  Für die Nutzung der Daten ist die  Datenlizenz Deutschland - Namensnennung - Version 2.0 anzuwenden. Die Lizenz ist über https://www.govdata.de/dl-de/by-2-0 abrufbar. Der Quellenvermerk gemäß (2) der Lizenz lautet "Geoportal Berlin / [Titel des Datensatzes]".
#> 6                                                                                                                                                                                                                               Es gelten keine Zugriffsbeschränkungen
#> 7                                                                                                                                                                                              https://fbinter.stadt-berlin.de/fb/wfs/data/senstadt/s_wfs_alkis_bezirk
#> 8                                                                                                                                                                                                                                                                  WFS
#> 9                                                                                                                                                                                                                                                                1.0.0
#> 10                                                                                                                                                                                                                                                               1.1.0
#> 11                                                                                                                                                                                                                                                               2.0.0
#> 12                                                                                                                                                                                                                                                     GetCapabilities
#> 13                                                                                                                                                                                                                                                 DescribeFeatureType
#> 14                                                                                                                                                                                                                                                          GetFeature
#> 15                                                                                                                                                                                                                                         text/xml; subtype=gml/2.1.2
#> 16                                                                                                                                                                                                                                         text/xml; subtype=gml/3.1.1
#> 17                                                                                                                                                                                                                                         text/xml; subtype=gml/3.2.1

### Multiple
ids_wfs <- readLines(system.file("extdata/ids_wfs.txt", 
package = "kwb.fisbroker"))
#> Warning: file("") only supports open = "w+" and open = "w+b": using the former

wfs_meta_list <- setNames(lapply(ids_wfs, function(id) {
kwb.fisbroker::read_metadata(id)}),
ids_wfs)

wfs_meta <- dplyr::bind_rows(wfs_meta_list, .id = "id_wfs")

wfs_meta
#> # A tibble: 0 × 0